used CORE_ANNOTIONS.HIDDEN to define INSTANCE_ANNOTAITONS.HIDDEN
